HTML Reporter: Avoid leaking 'Map' in older browsers
On IE6-8 and in PhantomJS, our partial ES6 Map implementation
was leaking as global variable.

Follows https://github.com/qunitjs/qunit/commit/2f2eaba36e9c4be7eaa82a71107ea5f66c7b1519.
Fixes https://github.com/qunitjs/qunit/issues/1508.
